Refactor: 인증 관련 도메인 모델 및 리포지토리 코어 모듈로 이동
- User 모델 클래스를 `feature/auth`에서 `core/domain`으로 이동
  - `Timestamp` 타입을 `Long`으로 변경하여 Firebase 의존성 제거
- AuthRepository 인터페이스를 `feature/auth`에서 `core/domain`으로 이동
  - `GoogleSignInAccount` 및 `FirebaseUser` 타입을 `Any`로 변경하여 특정 라이브러리 의존성 제거
  - DataStore 관련 함수 추가 (로그인 정보 저장/삭제/조회)
